Startsession metric fix (#16833)
## Description

We are occasionally having cases where StartSession metric is getting
logged twice for the same session. This is happening because we are
recreating a start session metric during some race conditions instead of
no-oping it.

This is an example case where this race condition is happening. Please
not the duration for the last extra startsession at the very end. It is
only around 0.022 ms


![image](https://github.com/microsoft/FluidFramework/assets/448106/69fba155-0e96-48a9-b2e3-2401d9334917)