tree2: update op size tests (#16792)
## Description

These tests were using the highest level tree API and mid level schema
API (which are getting removed. While its questionable able that they
use the full tree+runtime setup, if they are going to, they should use
the matching schema API, so that was changed.

Initialization was changed to use schematize.

Places calling jsonableTrees Json have been renamed: JsonableTrees are
not in the Json domain, nor are they json, and both of those are things
we do use and need to be clear arn't used here.

Byte lengths have been clarified to be length of utf8 encoding of
jsonable tree representation serialized to json. This doesn't seem like
the right thing to care about or measure, but its what it does and its
more clear now.

Top level functions now use the same syntax instead of a subset being
arrow functions.

Return types are now specified for all functions.

When building trees literals the children are constructed before the
parents to remove the need to modify the resulting tree.

expectChildrenValues has been renamed, documented and the implementation
cleaned up so the first child is no longer special cases.


Actual behavior of the tests should be unchanged.
This is just removing use of old APIs and clarifying things.