Turn off socket timeout by default (#16836)
we found out the production connection poll socket is keep on
close/reconnect with a very high rate. Looking at these, they were
controlled by this `DefaultSocketTimeoutMS` setting. By tracking the
source code, basically sdk will set the connection pool socket with this
as a socket timeout, and the socket connection will goes to closed state
with this setting.
https://nodejs.org/api/net.html#socketsettimeouttimeout-callback

However, not sure if the sdk had an issue of that, looks like ALL
connections are marking as error out state based on this setting on
production. This setting was turned off by SDK as set to 0 by default.
So just turn it off for now, we can inject customized settings.

We still have connection timeout and max idle timeout do the same thing.

Co-authored-by: Xin Zhang <zhangxin@microsoft.com>