Adjusting some of the perf tests to improve reliability numbers (#16825)
## Description

The reliability of the AFR runs is impacting our comparative analysis
against ODSP and the perf dashboard visualization over time. In order to
improve it, we are adjusting some of the tests to decrease the load when
running summaries, adding a little bit of delay between iterations and
increasing the default acktimeout for the tests. Also, in case of
exception, we want the summaries of other scenarios to run, so we are
adding a simple try catch to deal with these cases where for some reason
we time out.