Fix SharedString Event and Submit Order to Avoid Reentrancy (#16815)
This change primarily addresses a problem in sequence where the event
for a change is triggered before the op for the change is submitted.
This is a problem as if an application does reentrancy, specifically
creates more changes in the event for a change, it will lead to the
subsequent changes, made is the event, to be submitted before the
original change. This can result in document corruption.

We fix this issue by explicitly submitting before we fire the event.
Since we now submit on all events, it is also necessary to not submit in
some cases, specifically where there are stashed ops, as these may have
already been submitted.

Lastly, this change deprecates grouped ops, which also have reentrancy
concerns, specifically that if a change is made during the application
of a group op, that change cannot be correctly applied, as the group op
only has a single sequence number so there is no way to identify at
which sub-op the new change should be made after. Remote client will end
up processing the new change after the group op is complete, which is
not when the new change was generated, and can lead to eventual
consistency issues.

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>
Co-authored-by: Wes Carlson <49205066+wes-carlson@users.noreply.github.com>