Rename reason to summarizeReason so that it is not overwritten on summarize failure (#16828)
The `reason` for summarization is overwritten by [this
code](https://github.com/microsoft/FluidFramework/blob/277f5b8e89b4c0c3dcdac95fe4b5c7063cfdb870/packages/runtime/container-runtime/src/summary/summaryGenerator.ts#L293)
if summarization fails. This means that the ``Summarizer_cancel` event
will not have the reason for summarization. It would be helpful to know
the reason for the summarization that fails.

This PR renames the summarize reason to `summarizeReason` instead of
just `reason` which is generic and can be overwritten.