Reboot pods if it didn't meet min total threshold but failed consecutively (#16843)
Current pods rebooting for a meaningful number of request only. However
during weekend, pods might still stuck at db layer but not able to cure
due to it is not reach the meaningful number threshold. Adding this
consecutively counter to measure the case for low traffic time and we
failed consecutively we still refresh the pods

---------

Co-authored-by: Xin Zhang <zhangxin@microsoft.com>