Fix the seqnum issue when creating subdirectory with grouped batching - release 6.0 (#16857)
**Cherry-picking https://github.com/microsoft/FluidFramework/pull/16806
to the release branch. This is not intended to fix an ongoing bug, as
the feature (op grouping) is disabled. However, if enabled via options,
this bug will surface. Therefore, this PR is for prevention.**

## Description


[AB#5225](https://dev.azure.com/fluidframework/internal/_workitems/edit/5225)

The previous implementation lacks the capability to manage the case
where multiple `createSubDirectory` messages with grouped batching
enabled. For instance:

```typescript
directory.createSubDirectory("parent").createSubDirectory("child")
```

With grouped batching enabled, the message for creating the `/parent`
directory carries the same sequence number as the message for creating
`/parent/child`. However, the sequence number of `/parent/child` will be
-1 by default and cannot be updated with the previous logic. This
impacts the function `isMessageForCurrentInstanceOfSubDirectory` and may
lead to the defect of the function `processSetMessage`, which
subsequently causes the unexpected behavior when a remote client updates
value for `/parent/child`

Also include the test of
https://github.com/microsoft/FluidFramework/pull/16731 in this PR

Co-authored-by: clarenceli-msft <86381878+clarenceli-msft@users.noreply.github.com>
Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>