Add ContainerRuntimeMessage.compatDetails to futureproof against future "optional" ops (#16364)
Fixes
[AB#4416](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4416),
in preparation of adding GC "reference" and / or "deleted" ops -
[AB#4168](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4168).

## Background

Currently, if a new container runtime op type is added and sent, it will
crash older container runtimes because they will hit
[this](https://github.com/microsoft/FluidFramework/blob/b294c000fafcf7b374e13ac0b81605dc44fef829/packages/runtime/container-runtime/src/containerRuntime.ts#L2002)
assert. This makes adding new op types very time consuming. It has to be
done in 2 stages - First, ship code that can read the new op type and
once that saturates, ship code that writes the new op type.

## Description

This PR adds the ability for container runtime to not fail on receiving
unknown op type but rather treating it as no-ops. This will enable
adding new op types faster and it can be decided on a feature by feature
basis to whether to wait for saturation or not. For example, the GC
reference and deleted ops do not require saturation to be used.

@agarwal-navin is largely to thank for the design, which is to create a
new optional `compatDetails` object on `ContainerRuntimeMessage` (the
unpacked op). For now, it just has a `behavior` property which is either
`Ignore` or `FailToProcess`. But being an object (not the string alone),
it is extensible, in case variable data is needed to determine
compatibility for some future `behavior` type.