Server: Fix prettier warnings when building GitRest/Historian in container (#16234)
This prevents 'prettier' from emitting warnings for files in
'.pnpm-store/' when building the GitRest and Historian services per the
dev instructions in the README.md.

Because the dependencies required to build these services are tricky to
install locally (esp. on non-Linux systems), the README for these
services recommends:

* Building the service in a Docker container
* Mounting the 'gitrest/' or 'historian/' directories from the local
filesystem into the container.
* Running 'bash' inside the container to run 'pnpm i' and 'pnpm run
build'

A side effect or running 'pnpm i' with the local file system mount as
described above is that it places the '.pnpm-store' cache in the build
directory, causing 'prettier' to include this directory while scanning.