Deprecate getStackContext (#16877)
Mark the getStackContext APIs and ReferenceType.NestBegin/NestEnd (only
used for getStackContext) as deprecated. No one uses them and it causes
bookkeeping in hierBlock (rangeStacks). It's also not really tested.

The later removal is not yet scheduled (the functionality needs to be
removed from a couple of our examples) but marking them deprecated
should discourage new usages.