Safeguard against reentrancy on local op submission in sequence (#16810)
## Description

Adds a safeguard to prevent local changes to `SharedString` from
reentrant-ly incurring further local changes.
Before #16815, this basically always resulted in eventual consistency
issues or doc corruption and frequently crashes. #16815 fixes FF and
prevents it from corrupting documents, but our test coverage for this
type of thing is still not great.

We're guessing most SharedString users won't build models in a way
that's robust to reentrancy or want it anyway, so this adds checks which
fail fast by default when encountering reentrancy. Reentrancy can be
explicity opted into by setting `sharedStringPreventReentrancy` to
`false` on the data store options, see `reentrancy.spec.ts` in this
commit for more details and an example.