Tree: Use correct cell ID when referring to unattached cells (#16712)
## Description

Added a `detachIdOverride` field to sequence field marks which can
represent unattaches (Delete and ReturnFrom). When creating the inverse
of an attach this field is set to the attach's cell ID. We use this as
the output cell ID of the detach instead of its own ID and revision.

Also created an `adjacentCells` field on `CellId` which lists the IDs of
all adjacent cells detached in the same revision. This is used instead
of including those adjacent detaches in `CellId.lineage`. This
eliminates the need to handle splitting and merging of lineage event
when splitting and merging marks.

---------

Co-authored-by: yann-achard-MS <97201204+yann-achard-MS@users.noreply.github.com>