Deprecated refreshLatestAck summarize option (#16907)
The `refreshLatestAck` summarize option has been deprecated. This does
not work as expected because the container shuts down when summarization
is invoked with this option (See #15009 for details).
This will be removed from the summarize options in a later release.

This PR also does a couple other small refactors:
1. In `RunningSummarizer`, removed the `cancelationToken` params from
`trySummarize` methods. No one passes in this paramter; they use
`this.cancelationToken`.
2. Changed the style of `summarizeOnDemand` / `enqueuSummarize` params
and return types to make them easier to understand.