feat(check:policy): Add configurable policy for package names and scopes (#16863)
Previously, the list of scopes and package names we check in
policy-check was hardcoded. This change lifts those settings into the
fluid-build config, so it's easier to update if/when we need to change
policy.

The change won't take effect until the client release group is updated
to a version of build-tools that includes it, but I included the
fluid-build config changes in this PR, so they'll already be in place
when build-tools is upgraded.