build: Limit use of lerna in server CI pipelines (#16052)
We currently use lerna in some places in CI, but it is called via npx
and thus its dependencies are not locked to a particular version., This
means that lerna releases can break our pipelines. This PR adds lerna as
a root dependency to the server release groups, so that it and its
dependencies versions are pinned.

Despite the fact that lerna is still installed, I updated the scripts of
the release groups to use pnpm -r run instead of lerna. That reduces the
use of lerna so it will be easier to remove once we can.