Pass relevant override options for Container's detached/attach flow (#16889)
## Description

There is currently a gap in functionality when using the `load` flow vs
the `createDetached` flow on the Container. For the `createDetached`
flow, there was no way to provide override values for things like
`canReconnect` and `clientDetailsOverride` on the
`IContainerCreateProps` interface.

Now, the Loader's `createDetachedContainer(...)` and
`rehydrateDetachedContainerFromSnapshot(...)` methods contain an
optional props argument to pass in these values.

Also, for the `attach` method there was no way to configure connecting
to the deltaConnect (normally done through the `LoaderHeader.loadMode`
header in the `load` flow). Now, the Container's `attach(...)` method
contains an optional props argument to pass in this values.


[AB#4958](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4958)