Fix reconnection and rebase bugs in PropertyDDS (#16277)
## Description

This PR fixes three bugs in the PropertyDDS:
* When a client goes offline and reconnects later, it could happen that
messages are submitted which reference previous changes that are no
longer within the collaboration window have been pruned in the last
summary. With this patch, we will now submit a rebased version of these
old operations to make sure the reference operation is within the
collaboration window.
* There is a scenario, where the rebased local operations do not exactly
match the received remote operation. This can happen when changes in the
base cancel out (i.e. if there is a remove/insert combination). In this
case, if you rebase across the squased changeset within the remote
changes, a modification will be preserved. However, if you rebase
separately across remove and then insert a modification will be lost. We
did not correctly take this into account in the function that removes
local operations, once they have been confirmed from the server. We just
assumed that they will always be exactly the same, now we perform a
rebase if necessary.
* There was a bug in the rebased code, that did not correctly handle the
case when a modification was rebased with respect to a insert/remove
operation. This case was supposed to remove the modification (because it
was assumed to apply to the removed entry and might make no sense in the
new context any longer). Due to the bug, the modification was not
correctly removed.

## Breaking Changes

I modified the MockContainerRuntimeFactoryForReconnection to take into
account that a client that is no longer connected will not be considered
when computing the minimum sequence number on the server. I think this
behaviour is more realistic and at least in my case it was necessary to
reproduce the bug. Still, this will also affect other tests that are
based on this class and introduce this modified behaviour in those
tests. If this is not desired, we could make this optional.

A second aspect to consider here is, that this PR changes the behaviour
of the rebase for the third of the cases described above. So if multiple
clients on the same documetn are using different versions of the DDS,
they might get out of sync, since they resolve the rebases differently.
How should we handle such a situation?

---------

Co-authored-by: Mikhail Goncharov <17141437+sharptrip@users.noreply.github.com>
Co-authored-by: Daniel Lehenbauer <DLehenbauer@users.noreply.github.com>