feat(eslint-config-fluid): Enable import/no-deprecated (#16181)
Enables the
[import/no-deprecated](https://github.com/import-js/eslint-plugin-import/blob/main/docs/rules/no-deprecated.md)
lint rule in all our configs. I also bumped the config version to a new
minor version since 2.0.0 was already released.

I disabled it for test files, since they often need to import deprecated
stuff for test purposes.

This rule is useful as we refactor and move APIs to different packages,
because it helps us find places where we're using APIs that have been
replaced. This will be particularly useful with common-definitions and
common-utils.

In my testing, the following packages had legitimate failures with this
rule:

- @fluid-internal/test-version-utils
- @fluid-example/webflow
- @fluid-example/table-document
- @fluid-experimental/sequence-deprecated
- @fluid-experimental/attributable-map
- @fluidframework/map
- @fluidframework/merge-tree
- @fluidframework/container-runtime

However, the offenses can be addressed easily or ignored once this
change has made it into the client release group.