tree2: Add schema linting configuration (#16919)
## Description

Sometimes we need to make weird schema for testing or compatibility
purposes. This allows us to still inform people when they do it on
accident, but opt out of such checks as needed. This will also make
adding new checks in the future more acceptable.

Also removes unneeded SchemaCollection abstraction layer.

## Breaking Changes

Schema builder constructor now has an extra argument that must be
provided (default to {}) before the list of libraries if providing
libraries.