Remove old code path in token validation flow (#16954)
Previously in the PR:
https://github.com/microsoft/FluidFramework/pull/15612/files
I refactored the token validation flow to group the logic in one
function: verifyToken
For safety reason, it is guarded by this flag: enableTokenCache. this
flag enableTokenCache not only controls whether we use token cache or
not, but also controls whether we use the new validation function
(refactored function: verifyToken).

Token cache change has been deployed and validated in FRS. Now we don't
need to keep the old token validation flow anymore.

This PR just removes the old code path in token validation while keep
the logic same as before. The removed logic should already be covered in
verifyToken function, regardless of the enableTokenCache flag.

Within the enableTokenCache, it has all the logics to validate a token:
- valid claims
- token expired or not
- token revoked or not
- single token check if enabled
- use cached token if enabled
- send request to riddler for token validation if cache is not enabled
or token not found in cache.

Co-authored-by: Yunpeng Hou <yunho-microsoft@Yunho-MacBookPro2019.local>