GC: Add option to force loading an Inactive object to fail, to support partners instrumenting GC-impacting bugs (#16874)
Applications would like to know if an object they request is inactive or
tombstoned. They'll use this to instrument their key object load
codepaths to get a clear picture of whether there's any indication of
GC-blocking bugs in the system.

Tombstone already supports this - the error thrown when accessing a
Tombstoned object has a particular flag on it that indicates as such,
and you can call `Container.resolveHandle` passing a header to bypass
the Tombstone failure.

This commit adds the same logic for Inactive objects.