refactor: Move loader-agnostic error types from `container-utils` to `telemetry-utils` (#16895)
This is part of an effort to remove the container-utils package. The
objective is to move the error types in that package to
`telemetry-interfaces`. This PR moves all of the loader-agnostic types
to their new home.

The moved members are re-exported from `telemetry-interfaces` to
maintain backwards compatibility but are marked as deprecated.
Repo-internal imports have been updated to import from
`telemetry-interfaces`, and dependencies on `container-utils` were
removed where possible.

Types moved and their destinations:
* DataCorruptionError: Moved to `telemetry-utils`, re-exported and
deprecated in `container-utils`
* GenericError: Moved to `telemetry-utils`, re-exported and deprecated
in `container-utils`
* UsageError: Moved to `telemetry-utils`, re-exported and deprecated in
`container-utils`
* ThrottlingWarning: Copied to `container-loader` (not exported),
deprecated in `container-utils`