refactor: Move ClientSessionExpiredError from `container-utils` to `container-runtime` (#16964)
ClientSessionExpiredError is exclusively used by `container-runtime`.
This PR moves a copy of the type from `container-utils` (which is
scheduled for removal) to `container-runtime`, where it is no longer
exported. The existing copy in `container-utils` has been marked as
deprecated.

Also adds a `@packageDocumentation` comment to `container-utils` to
notify consumers that the package itself is considered deprecated.