Add verbose logging to the telemetry system for stress tests (#16845)
## Description

Just tried a sample implementation. Not sure if it is the direction
which is digestible or not. Please review.
Add general implementation for verbose logging and use it for service
load tests. Added config to base logger so that anybody who is supplying
this base logger can specify what logs needs to send to the telemetry
system. For now, I just needed verbose logging, so filtered out on that.
This can be used/modified to filter out based on namespace etc.
Table for category/loglevel intersection
<p>&nbsp;</p><figure class="table">

Category/LogLevel | Verbose | Default | Warning | Error
-- | -- | -- | -- | --
Performance | 1 | 1 | x | x
Generic | 1 | 1 | 1 | x
Error | x | x | x | 1

</figure><p>&nbsp;</p>

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>