Tree: Add test demonstrating sequence rebasing bug (#16967)
## Description

Currently when rebasing a sequence mark targeting empty cells over a new
insert we assume that the new insert's cells are newer than the empty
cells, but this is not true if the empty cells were created on the
branch we're rebasing. This could be fixed by providing the rebaser the
relative order of the cell revisions, or by avoiding sandwich rebasing
altogether. This PR adds a comment in the relevant code and a test
demonstrating the issue.