fix(fluid-build): fix incremental builds for TS 5.1 (#16985)
The logic in 'fluid-build' that determines when to recompile a
TypeScript project includes enumerating the files and their hashes
listed in "*.tsbuildinfo" and comparing these with the file's current
hash.

In previous versions of TypeScript, the hash stored in *.tsbuildinfo was
just the sha256 has of the file contents. Newer versions of the
TypeScript compiler preprocess the source file contents to strip the
sourcemap comment at the end of the file (if any), which alters the
resulting sha256 hash.

This change alters the TscTask to instead call into the TypeScript
compiler's internals to compute the hash.