Emit summarize event on every summarize event for better observability (#17003)
## Description
Added emitting a `summarize` event whenever a summarize attempt
completes in `RunningSummarizer`. This event is propagated all the way
to the summarizer's parent client where it is emitted by the container
runtime. This should provide better observability into summarization and
also help us write tests.
Added minimal properties when the event is generated. More properties
can be added as and when required.

## Why is this needed
Currently, there is no way today to know when a summarization attempt
fails. Summarization success can still be known by waiting for a summary
op or ack. This makes it especially hard to write tests where we want to
validate summarization attempt failures, retries, etc. For example,
[this PR](https://github.com/microsoft/FluidFramework/pull/16961) adds
the capability for summarization to succeed after 5 retries. There is no
easy way to test this other than adding some nasty hacks.
The other thing that makes this hard is that there is no way to get hold
of the summarizer client unless one is explicitly created. So, testing
the states of heuristic based summarization is really hard. So, the
"summarize" events have to be emitted by the parent client as well.


[AB#5349](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5349)