Fix containers created via clone not added to LoaderContainerTracker (#16987)
## Description
`LoaderContainerTracker` tracks all containers created via loaders that
are created by TestObjectProvider. An expection to this is summarizer
containers which are only tracked if `syncSummarizerClients` option is
true.
It does so by overriding the container creation methods in `Loader` and
adding those containers to a list. However, `Container` has a `clone`
method that can be used to create a new container without going through
the `Loader` apis. For instance, summarizer containers are create via
the `clone` API.
This change adds tracking of containers that are created via the `clone`
API.

## How I found the issue
I found this issue when writing a test with `syncSummarizerClients`
option as true and waiting for `ensureSynchronized` after summarizer
client spawned. `ensureSynchronized` never completes because it doesn't
track summarizer container as opened but the `getPendingClients` method
sees the summarizer container in quorum and thinks its supposed to leave
and so, it waits for the `removeMember` event which never happens.