Call container dispose on load failure (#17013)
## Description

Currently we only call container close on load failure, but it does not
free up all resources. Also, the app does not have the container at this
stage to call dispose on it. So, it should be called on load failure.

This issue was caught when adding test to check whether ODSP cache was
cleared on load failure with error or not. So, adding that test and also
modify the other test to look for container dispose event too.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>