tree2: Tidy op size tests (#16853)
## Description

Initially this change attempted to keep the test results exactly the
same. However after realizing that before this change there were extra
ops getting counted that shouldn't exist, it was pivoted to cleaning up
that as well. As a result the data has change at the same time as the
test setup, which is not ideal, but given that these test's extra ops
issue existed when they were new (#13356), and changed unnoticed to be
worse sometime between then and now (up to 4 ops for tests that should
have 1), I think it makes sense to just to all this as once.


Fixes several style issues with the op size tests:
- Redundant string formatting. Every test name (which is a lot of
formatted text) occurred twice. Instead quert the info from the Mocha
Context.
- Move some test setup into the beforeEach hook to simplify the tests
(including the above).
- Avoid restating the whole hierarchy of each block in tests at every
level. instead use fullTitle to include the parent describe blocks.
- Use Enums instead of string intersections where appropriate.
- Stop prefixing all the tests with letter.number.letter etc. Just use
the descriptive parts of the names
- Share the utility to compute the size of each message instead of
reimplementing it in multiple places
- Compute the size of each message once, instead of multiple times for
each stat.
- Replace foreEach loops with .map,  and .reduce where appropriate.
- Use max(...numbers) to compute max instead of a forEach loop.
- Remove special case for no data (this case now would work without
special cases and also is an error).
- Use table driven patterns to generate some of the tests to loops
instead of hand written full cross products.

Functional changes:
- Make tests error if no data is collected
- Adjust test setup to avoid depending on mock services (works with
purely local edits).
- No longer gets 4 ops for the single op cases (likely due to old
version double counting all ops due to sequencing in processMessages,
plus some other unknown issue)
- No longer reports average message or min size as they are not very
useful. Instead add total size.
- Tests results are no longer sorted by average op size: this made
comparing before and after results needless difficult and put everything
in an order that made it hard to find things.


Also:
- Add documentation covering limitations of the tests

Not changed:
- What cases are tested and what edits are performed in them.
- Some of the tests weren't updated to the table driven pattern, mostly
due to my lazyness and wondering if we even want to keep those.

This is a continuation of #16792 which cleaned up the helpers in this
file.