Fixed reconnect and rebase bug in PropertyDDS. (#17014)
## Description

This is a backport of the PR #16277 to the lts branch.

It three bugs in the PropertyDDS:

* When a client goes offline and reconnects later, it could happen that
messages are submitted which reference previous changes that are no
longer within the collaboration window have been pruned in the last
summary. With this patch, we will now submit a rebased version of these
old operations to make sure the reference operation is within the
collaboration window.
* There is a scenario, where the rebased local operations do not exactly
match the received remote operation. This can happen when changes in the
base cancel out (i.e. if there is a remove/insert combination). In this
case, if you rebase across the squased changeset within the remote
changes, a modification will be preserved. However, if you rebase
separately across remove and then insert a modification will be lost. We
did not correctly take this into account in the function that removes
local operations, once they have been confirmed from the server. We just
assumed that they will always be exactly the same, now we perform a
rebase if necessary.
* There was a bug in the rebased code, that did not correctly handle the
case when a modification was rebased with respect to a insert/remove
operation. This case was supposed to remove the modification (because it
was assumed to apply to the removed entry and might make no sense in the
new context any longer). Due to the bug, the modification was not
correctly removed.