Implement Package Promotion in ADO Build Feed (#16610)
### Objective:
Automate the promotion of the release packages in ADO's "Build" feed to
the Release view, ensuring their indefinite retention.

### Changes:
- When the pipeline publishes the package, detect if its a release in
the "Build" feed.
- Promotes these packages to the Release view in Azure Artifacts.

### Testing;
- I have tested it using the SystemAccess Token provided in the
pipelines;
https://dev.azure.com/fluidframework/internal/_build/results?buildId=185450&view=results
- running a prerelease using my PAT with the url pointing to the
testRishhiRetentionPolicy feed instead of the builld feed, seeing that
package get promoted in its feed
(https://pkgs.dev.azure.com/fluidframework/internal/_apis/packaging/feeds/testRishhiRetentionPolicy/npm/@fluidframework/agent-scheduler/versions/2.0.0-internal.5.2.0?api-version=7.1-preview.1);
https://dev.azure.com/fluidframework/internal/_build/results?buildId=185432&view=results
- Locally promoting specific versions and watching if retention policy
deletes those packages, which it did not

We get $packageName in the bash script in the following format;
fluidframework-azure-client. To promote this package we need to convert
it back to the official package name as found in the package.json. For
this reason I have the following string manipulation;
https://github.com/microsoft/FluidFramework/blob/115db04715e3f695db4ca9d1d324dcdd6630e497/tools/pipelines/templates/include-publish-npm-package-steps.yml#L146-L153