Driver changes to support ephemeral containers #Draft (#16006)
### Description

Ephemeral containers are containers whose artifacts (ops, summaries and
metadata) are stored only for the lifetime of the session. The container
lifecycle is completely controlled by the fluid server (for durable
containers, the deletes are client initiated). We make use of the
comparatively shorter lifetime of these containers to store their
artifacts closer to the user without breaking compliance which helps
these containers have better latency and reliability while using less
cogs.

The flow is initiated by the client that sends the isEphemeral container
as a part of the create document flow