Remove connection exceptions from a few end-to-end tests (#17052)
## Description

ADO:4131
ADO:3986
ADO:3983

In preparation for switching to 'read' connection by default, these
tests need to be 'fixed' to accommodate for that change. The tests (and
what they are testing) are oblivious to the connection defaults, as they
should be, but they do have dependencies on the side effect of the
connection mode (default being 'write'), specifically the order of the
ops and/or whether a summary is created before the test starts.