Upgrade pipelines to Node18 (#17077)
Update the recommended versions of Node.js for developers to use, as
well as the version used by CI.

This updates from version 16 (which goes out of support on 2023-09-11)
to version 18 (which goes out of support on 2025-04-30).

The minimum, as enforced by the root package.json's "engines" block is
not updated to 18 yet. This can be done after a migration period.

This has no impact on the produced packages, but will allow (but not yet
require) developers working in this repo to upgrade.