Remove refresh from server path and refresh from network call (#16657)
[AB#5118](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5118)

Remove "refresh from server" logic. We are making this PR into `next`
instead of `main` as we want to wait a bit before we commit to removing
"refresh from server" logic. Instead of refreshing we are closing the
container and restarting. More details here:
https://github.com/microsoft/FluidFramework/pull/15140

This requires updating the `summarizerNode` logic and `containerRuntime`
logic. Work to move when we decide to close the container sits in
[AB#5152](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5152)

---------

Co-authored-by: Navin Agarwal <45832642+agarwal-navin@users.noreply.github.com>
Co-authored-by: Mark Fields <markfields@users.noreply.github.com>